+++ pkgs/test_core/lib/src/runner/configuration.dart
@@ -43,6 +43,9 @@ class Configuration {
   bool get help => _help ?? false;
   final bool _help;
 
+  /// Custom HTML template file.
+  final String customHtmlTemplatePath;
+
   /// Whether `--version` was passed.
   bool get version => _version ?? false;
   final bool _version;
@@ -216,6 +219,7 @@ class Configuration {
 
   factory Configuration(
       {bool help,
+      String customHtmlTemplatePath,
       bool version,
       bool pauseAfterLoad,
       bool debug,
@@ -263,6 +267,7 @@ class Configuration {
     var chosenPresetSet = chosenPresets?.toSet();
     var configuration = Configuration._(
         help: help,
+        customHtmlTemplatePath: customHtmlTemplatePath,
         version: version,
         pauseAfterLoad: pauseAfterLoad,
         debug: debug,
@@ -323,6 +328,7 @@ class Configuration {
   /// Unlike [new Configuration], this assumes [presets] is already resolved.
   Configuration._(
       {bool help,
+      String customHtmlTemplatePath,
       bool version,
       bool pauseAfterLoad,
       bool debug,
@@ -346,6 +352,7 @@ class Configuration {
       bool noRetry,
       SuiteConfiguration suiteDefaults})
       : _help = help,
+        customHtmlTemplatePath = customHtmlTemplatePath,
         _version = version,
         _pauseAfterLoad = pauseAfterLoad,
         _debug = debug,
@@ -470,6 +477,8 @@ class Configuration {
 
     var result = Configuration._(
         help: other._help ?? _help,
+        customHtmlTemplatePath:
+            other.customHtmlTemplatePath ?? customHtmlTemplatePath,
         version: other._version ?? _version,
         pauseAfterLoad: other._pauseAfterLoad ?? _pauseAfterLoad,
         color: other._color ?? _color,
@@ -511,6 +520,7 @@ class Configuration {
   /// always replaced by the new one.
   Configuration change(
       {bool help,
+      String customHtmlTemplatePath,
       bool version,
       bool pauseAfterLoad,
       bool color,
@@ -554,6 +564,8 @@ class Configuration {
       Iterable<String> addTags}) {
     var config = Configuration._(
         help: help ?? _help,
+        customHtmlTemplatePath:
+            customHtmlTemplatePath ?? this.customHtmlTemplatePath,
         version: version ?? _version,
         pauseAfterLoad: pauseAfterLoad ?? _pauseAfterLoad,
         color: color ?? _color,